Mon, Jul 4, 2022, 12:46 PM +00:00

+++ pset4/recover/recover.c
@@ -5,7 +5,7 @@
 
 typedef uint8_t BYTE;
 
-int BLOCK_SIZE = 512;
+BYTE BLOCK_SIZE = 512;
 
 int main(int argc, char *argv[])
 {
